Вынесены модели данных за пределы класса MovieQuizViewController. В контроллере MovieQuizViewController применил модификатор weak для аутлетов (слабые ссылки на элементы интерфейса).  Проверку наличия данных и декодирование JSON, а также получение пути к директории документов вынесены из метода viwDidLoad. Изменена структура GameRecord, в которой применено расширение Date+Extensions, проведена замена метода recordText на свойство recordText.
