+++ MovieQuiz/Models/ActorModel.swift
@@ -1,0 +1,14 @@
+//
+//  ActorModel.swift
+//  MovieQuiz
+//
+//  Created by Чингиз Джабаев on 26.07.2023.
+//
+
+import Foundation
+struct Actor: Codable {
+    let id: String
+    let image: String
+    let name: String
+    let asCharacter: String
+}
