+++ MovieQuiz/Models/ViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ViewModel.swift
+//  MovieQuiz
+//
+//  Created by Чингиз Джабаев on 26.07.2023.
+//
+
+import UIKit
+struct ViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
